// frontend/src/payload-types.ts
export interface ChartBlock {
  chartType: 'mermaid' | 'bar';
  title?: string | null;
  caption?: string | null;
  mermaidSource?: string | null;
  barData?:
    | {
        label: string;
        value: number;
        highlight?: boolean | null;
        id?: string | null;
      }[]
    | null;
  yLabel?: string | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'chart';
}